+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 ## Upcoming...
 
 - ... <!-- Add new lines here. Version number will be decided later -->
+
+## 0.4.0
+
+_2020-11-11_
+
 - **BREAKING** feat: Make `useCallbackOnEsc` accept `deps` array (default:
   `[]`)
 - **BREAKING** feat: Revert `useCallbackOnEsc` to "run on mount" by default